tests/wpt/tests/css/css-gaps/parsing/rule-outset-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: *-rule-outset getComputedStyle()</title>
  <link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#outset">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <style>
    #target {
      font-size: 40px;
    }
  </style>
  <script>
    test_computed_value("column-rule-outset", "10px");
    test_computed_value("column-rule-outset", "-20px");
    test_computed_value("column-rule-outset", "0.5em", "20px");
    test_computed_value("column-rule-outset", "calc(10px + 0.5em)", "30px");
    test_computed_value("row-rule-outset", "calc(10px - 0.5em)", "-10px");
    test_computed_value("row-rule-outset", "30%");
    test_computed_value("row-rule-outset", "calc(25% + 10px)");
  </script>
</body>
</html>
